fix src lib not being used by prebuilt stage3 binary
The Zig lib dir is searched relative to the realpath of the Zig binary.

Because the Zig binary was called via a symlink, the Zig lib dir
in the build dir was not picked up and the Zig lib dir in the Nix
store was used instead.

Fix this by copying the Zig binary instead of symlinking.
In this way, the prebuilt Zig stage3 binary works exactly like the
bootstrap stage3 binary.